build 2.0.1 (#2)
In the future, we shouldn't "go back" like this with the LATEST file. We
should instead branch out from the latest commit that had that version.

For this transitory period, though, we don't really have that option.

We need a 2.0.1 folder on the S3 bucket in order for the release process
to work, because it will try to move 2.0.1 to the root as soon as 2.0.1
is marked as stable (i.e. the `prerelease` flag is removed).