fix PDF ts links
Reported by Ben on Slack: in the current PDF, the links are dead
(because they are relative links). This PR changes the build script to
turn them into absolute links in-between building the HTML and PDF
versions.

This isn't "clean" or "robust" by any means, but I can't think of any
other way to solve this.