[update root] move current version to 2.6.0 subdir
In order to be able to update the root website, we need to be able to
build the root website. This means that we need, at a minimum, two
separate versions of the docs editable and buildable at the same time:
the "next" version, on which most of the work happens, and the "current"
version, on which we may want to make occasional adjustments.

Right now the "next" version is 2.6.0 and the "current" one is 2.5.1.

However, giving the docs repo the ability to update root essentially
forces it to take over root updates generally, which means we'll benefit
from being able to build more than just the last two versions. More on
that later.

For the time being, this PR moves the next version under a subfolder;
the current version will be added as a subsequent PR, and then (as yet
another PR) I'll add the root update itself.

For most of the work, which touches only the next version, this should
be a very minimal disruption as all it changes is the work happens in a
nested folder, but everything stays the same.