fix pipfile issue
The `live-preview` script was using pipenv instead of the virtualenv we
now have, resulting in a second round of installing all the plugins and
the generation of the `Pipfile` and `Pipfile.lock` files.

Also fix a small possible issue where Sphinx would get the wrong version
of docutils from nix.