fix snapshots (#239)
There were two issues with the snapshot upload script that just
surfaced:

1. The snapshots were processed in reverse order (most recent first).
2. The script was building them but not actually pushing them.

Combined with the breaking change in Canton docs 0407 (see #238), this
lead us to a state where the most recent snapshot builds (without being
pushed), then the next-most-recent snapshot doesn't and the script fails.

This is the first step in addressing that: we fix the two issues above
and revert #238. With this change, the snapshot publishing script will
(hopefully) correctly push the previous snapshot then fail on the
current one, and then we can apply #238 again for the latest snapshot.

Apologies for the mess.