Replace canton.io references in ethereum.rst
Instead of using RST cross-referencing links, hyperlinks to the
documentation published on canton.io have been used.

As the documentation on canton.io is discontinued with the launch
of the Canton Network these references have to be replaced.

This commit also uses an updated canton tarball (`versions.json`)
which uses Sphinx relative paths for the scaladoc links.

This should make it possible to move the Canton Console Commands
and Error Codes reference pages to different directory locations
without breaking the therein contained scaladoc links.