Fix unnecessary warning when starting the monitoring
When running the monitoring example with `./dc.sh up -d` then the error
`network monitoring declared as external, but could not be found` is
reported.

This change defines the "monitoring network" as non-external which
prevents this error.

Using `external: true` was helpful during the development only, so that
docker-compose could be run many times, and adding new nodes to the
network.