add kubernetes section (#456)
* add kubernetes section

Add a section under Deploy Daml for a guide on Kubernetes. This commit is to convert existing files into the docs.daml.com repo. Later, a language review will need to be done.

* Apply suggestions from code review

Co-authored-by: carrielaben-da <91496516+carrielaben-da@users.noreply.github.com>

* address PR feedback

Updated terminology in diagram, added some download links, removed a couple sections as requested, and made a couple formatting changes.

---------

Co-authored-by: carrielaben-da <91496516+carrielaben-da@users.noreply.github.com>