snapshots: only look at first 2 numbers (#482)
Right now, in order to build a snapshot, the system looks at available
versions for a matching "stable": in order to build `2.8.0-snapshot`, it
would look for a `2.8.0` folder. That works for future stable versions
that aren't published yet, like 2.8.0, but it doesn't work for, say,
2.7.2-rc1, which will exist before we are ready to make a 2.7.2 folder.

With this change, we're only trying to match the first two segments,
i.e. the system would build `2.7.2-rc1` based on the `2.7.1` folder.