Update section 'Persistence using Postgres' (Canton Installation)
To avoid out-dated information, replaces the recommendation for a
specific Postgres version with a general statement and a link to
find the Canton tested Postgres versions.

Also asserts that the code snippets are just there for illustrative
purposes and may need to be adapted by the reader, e.g. choosing a
different Postgres Docker image.

Co-authored-by: marciw-da <marci.windsheimer@digitalasset.com>