Remove migration and replication from docs
There are breaking changes that make this section obsolete for Daml 3. Right now,
to make sure the docs team can freely bump to the latest Canton version freely,
we are simply removing the docs. We already scheduled the work necessary to
introduced the overhauled party migration and replication process once that is
fully sketched out and implemented.