Rewrote installation guide to use new reference config (#583)
Before, the installation guide was based on the advanced config example,
which was guiding people through the process by starting with the simplest
example and then calling out options.

Now, we start with a full example that comes with TLS, JWT, Storage
etc, where a client can then decide to incrementally remove configurations instead.

The docs changes are related to https://github.com/DACH-NY/canton/pull/15882