reorganize manual install page
The way the page is currently written assumes that the only reason you
might want to do a manual install is because you want to take the time
to check the signature.

What seems to be a more prevalent case based on my interactions with
Daml users is a need to install from a tarball on a machine that does
not have internet access. Right now, that use-case requires reading the
whole page and figuring out which steps you don't need to perform. It
also makes the manual install look very scary at first glance.

In this PR, I have extracted (and slightly supplemented) the "just the
install" steps to their own section, and left the signature verification
steps largely untouched. I expect this is far from perfect and this page
still needs a thorough pass by someone actually qualified to write
technical documentation, but I believe this will make things quite a bit
easier for our users in the short term.