Fix `literalinclude` of `IouMain.java` for the `quickstart.rst`
Some `literalinclude` refer to line numbers in `IouMain.java`, and have become
out of sync because the source code changed by one line.

This broke the build because on `literalinclude` specified a non-existent line
number.

This fixes the off-by-one problem.

The actual fix would be to use code comment markers instead of line numbers, but
that requires changing the source code and creating a new Daml artifact. So
that's maybe for another PR.