Bump 2.10.0 doc dependencies (#853)
* Bump Daml, Canton and Driver 2.10.0 doc dependencies

The 2.10.0 release happens soonish, so it's worthwhile to bump these dependencies.

* Fix `literalinclude` of `IouMain.java` for the `quickstart.rst`

Some `literalinclude` refer to line numbers in `IouMain.java`, and have become
out of sync because the source code changed by one line.

This broke the build because on `literalinclude` specified a non-existent line
number.

This fixes the off-by-one problem.

The actual fix would be to use code comment markers instead of line numbers, but
that requires changing the source code and creating a new Daml artifact. So
that's maybe for another PR.

* Fix `literalinclude` for Daml intro (9) 'Work with Dependencies'

The `literalinclude` marker does not exist anymore.

This is a quick-fix. The better fix would be use comments as markers rather than
actual line content.